requirements.txt
azure-functions
requests
azure-cosmos
